New HTTP fetcher, online portrait fetching (beta)
git-svn-id: svn://gtkevemon.battleclinic.com/GTKEVEMon/trunk/gtkevemon@4 6e31b030-b45c-3146-9160-cf88ccb029a2

+++ src/server.cc
@@ -14,7 +14,7 @@ int Server::current_socket = 0;
 
 /* ---------------------------------------------------------------- */
 
-Server::Server (std::string const& name, std::string const& host, int port)
+Server::Server (std::string const& name, std::string const& host, uint16_t port)
 {
   this->name = name;
   this->host = host;
